// ТРПИ/LP_Lab10_11/lab11/lab11/lab9/src/assets/models/IVacation.ts
export default interface IVacation {
    id?: number;
    img?: string;
    company?: string;
    name?: string;
    city?: string;
    lans?: string[];
    date?: string;
    grafic?: string;
    loud?: string;
    exp?: string;
    pay?: string;
    desc?: string[];
    expIn?: string[];
    isSelected?: boolean;
    onClick?: any;
    isLiked?: boolean;
}